refactor: split home page into authPage and StartInterivew components
